requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==1.3.5
python-socketio==5.10.0
SpeechRecognition==3.10.0
gTTS==2.4.0
pyaudio==0.2.13
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
uvicorn==0.24.0.post1
python-multipart==0.0.6
websockets==12.0